requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
setuptools==68.2.2
wheel==0.41.2
